AiGymChatBot/requirements.txt
langchain_core
python_dotenv
streamlit
langchain_community
langserve